Refactor and extend models with proper constructor and interface.
