--- frontend/angular/src/app/_models/ayah.model.spec.ts
@@ -1,7 +1,0 @@
-import { Ayah } from './ayah.model';
-
-describe('Ayah', () => {
-  it('should create an instance', () => {
-    expect(new Ayah()).toBeTruthy();
-  });
-});
