--- frontend/angular/src/app/_models/ayah.model.ts
@@ -1,4 +1,0 @@
-export class Ayah {
-    number: number;
-    text: string;
-}
