+++ frontend/angular/src/app/_models/deserializable.model.ts
@@ -1,0 +1,3 @@
+export interface Deserializable {
+    deserialize(input: any): this;
+}
