--- frontend/angular/src/app/_models/recitation.model.spec.ts
@@ -1,7 +1,0 @@
-import { Recitation } from './recitation.model';
-
-describe('Recitation', () => {
-  it('should create an instance', () => {
-    expect(new Recitation()).toBeTruthy();
-  });
-});
