--- frontend/angular/src/app/_models/recitation.model.ts
@@ -1,7 +1,0 @@
-export class Recitation {
-    surah: number;
-    ayah: number;
-    segments: string;
-    audio: string;
-    reciter: number;
-}
