--- frontend/angular/src/app/_models/reciter.model.spec.ts
@@ -1,7 +1,0 @@
-import { Reciter } from './reciter.model';
-
-describe('Reciter', () => {
-  it('should create an instance', () => {
-    expect(new Reciter()).toBeTruthy();
-  });
-});
