--- frontend/angular/src/app/_models/reciter.model.ts
@@ -1,7 +1,0 @@
-export class Reciter {
-    id: number;
-    name: string;
-    bitrate: number;
-    style: string;
-    slug: string;
-}
