--- frontend/angular/src/app/_models/surah.model.spec.ts
@@ -1,7 +1,0 @@
-import { Surah } from './surah.model';
-
-describe('Surah', () => {
-  it('should create an instance', () => {
-    expect(new Surah()).toBeTruthy();
-  });
-});
