--- frontend/angular/src/app/_models/surah.model.ts
@@ -1,14 +1,0 @@
-import { Ayah } from './ayah.model';
-
-export class Surah {
-    number: number;
-    ayahs: Array<Ayah>;
-    name: string;
-    total_ayahs: number;
-}
-
-export class SurahMain {
-    number: number;
-    name: string;
-    total_ayahs: number;
-}
